+++ engine/init.php
@@ -7,5 +7,18 @@
 
 use Core\Kernel;
 
+register_shutdown_function(function() {
+    try {
+        $data = \Dev\Collector::finish();
+        $cfg = \Dev\Logger::cfg();
+        if (($cfg['enabled'] ?? false) && ($cfg['log_requests'] ?? false)) {
+            \Dev\Logger::write('requests.log', ($data['request']['method'] ?? '') . ' ' . ($data['request']['uri'] ?? ''));
+        }
+        if (\Dev\DebugPanel::shouldShow()) {
+            \Dev\DebugPanel::render($data);
+        }
+    } catch (\Throwable $e) {}
+});
+
 $kernel = new Kernel();
 $kernel->boot();