+++ system/dev.php
@@ -1,0 +1,9 @@
+<?php
+return [
+  'enabled' => true,
+  'debug_panel' => true,
+  'log_requests' => true,
+  'log_templates' => true,
+  'log_sql' => true,
+  'storage' => ROOT_PATH . '/storage/dev',
+];